backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
werkzeug==3.0.1
bcrypt==4.1.2
